+++ app/src/main/java/ru/pavelcoder/modulbankdemo/model/retrofit/CurrencyRatesResponse.kt
@@ -3,5 +3,5 @@ package ru.pavelcoder.modulbankdemo.model.retrofit
 class CurrencyRatesResponse {
     val date: String? = null
     val base: String? = null
-    val rates: HashMap<String, Float>? = null
+    val rates: HashMap<String, Double>? = null
 }
